% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomparison.R
\name{detectNested}
\alias{detectNested}
\title{Discern nesting pattern of SEM coefficients}
\usage{
detectNested(models)
}
\arguments{
\item{models}{A List of lavaan-fitted SEM models}
}
\value{
matrix indicating nesting relationships
}
\description{
Receives a list of models and orders them by best guess at intended nesting
}
\author{
Ben Kite <bakite@ku.edu>
}
