% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomparison.R
\name{print.kutable}
\alias{print.kutable}
\title{A print method for kutable objects}
\usage{
\method{print}{kutable}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{optional arguments, corrently ignored}
}
\value{
x unchanged
}
\description{
A print method for kutable objects
}
